move-adb-stuf-to-provider
